Raise error if atlas affines don't match across participants (#1075)
