+++ xcp_d/tests/test_utils_atlas.py
@@ -45,57 +45,3 @@ def test_get_atlas_cifti():
 
     with pytest.raises(FileNotFoundError, match="DNE"):
         atlas.get_atlas_cifti("tofail")
-
-
-def test_copy_atlas(tmp_path_factory):
-    """Test xcp_d.utils.atlas.copy_atlas."""
-    tmpdir = tmp_path_factory.mktemp("test_copy_atlas")
-    os.makedirs(os.path.join(tmpdir, "xcp_d"), exist_ok=True)
-
-    # NIfTI
-    atlas_file, _, _ = atlas.get_atlas_nifti("Gordon")
-    name_source = "sub-01_task-A_run-01_space-MNI152NLin2009cAsym_res-2_desc-z_bold.nii.gz"
-    out_file = atlas.copy_atlas(
-        name_source=name_source, in_file=atlas_file, output_dir=tmpdir, atlas="Y"
-    )
-    assert os.path.isfile(out_file)
-    assert os.path.basename(out_file) == "space-MNI152NLin2009cAsym_atlas-Y_res-2_dseg.nii.gz"
-
-    # CIFTI
-    atlas_file, atlas_labels_file, atlas_metadata_file = atlas.get_atlas_cifti("Gordon")
-    name_source = "sub-01_task-imagery_run-01_space-fsLR_den-91k_desc-denoised_bold.dtseries.nii"
-    out_file = atlas.copy_atlas(
-        name_source=name_source, in_file=atlas_file, output_dir=tmpdir, atlas="Y"
-    )
-    assert os.path.isfile(out_file)
-    assert os.path.basename(out_file) == "space-fsLR_atlas-Y_den-91k_dseg.dlabel.nii"
-
-    # TSV
-    name_source = "sub-01_task-imagery_run-01_space-fsLR_den-91k_desc-denoised_bold.dtseries.nii"
-    out_file = atlas.copy_atlas(
-        name_source=name_source, in_file=atlas_labels_file, output_dir=tmpdir, atlas="Y"
-    )
-    assert os.path.isfile(out_file)
-    assert os.path.basename(out_file) == "atlas-Y_dseg.tsv"
-
-    # JSON
-    name_source = "sub-01_task-imagery_run-01_space-fsLR_den-91k_desc-denoised_bold.dtseries.nii"
-    out_file = atlas.copy_atlas(
-        name_source=name_source, in_file=atlas_metadata_file, output_dir=tmpdir, atlas="Y"
-    )
-    assert os.path.isfile(out_file)
-    assert os.path.basename(out_file) == "atlas-Y_dseg.json"
-
-    # Ensure that out_file isn't overwritten if it already exists
-    fake_in_file = os.path.join(tmpdir, "fake.json")
-    with open(fake_in_file, "w") as fo:
-        fo.write("fake")
-
-    out_file = atlas.copy_atlas(
-        name_source=name_source, in_file=fake_in_file, output_dir=tmpdir, atlas="Y"
-    )
-    assert os.path.isfile(out_file)
-    assert os.path.basename(out_file) == "atlas-Y_dseg.json"
-    # The file should not be overwritten, so the contents shouldn't be "fake"
-    with open(out_file, "r") as fo:
-        assert fo.read() != "fake"
